\name{RSSReQTL-package}
\alias{RSSReQTL-package}
\alias{RSSReQTL}
\docType{package}
\title{
\packageTitle{RSSReQTL}
}
\description{
\packageDescription{RSSReQTL}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RSSReQTL}
\packageIndices{RSSReQTL}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{RSSReQTL}

Maintainer: \packageMaintainer{RSSReQTL}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
